[P2] Rebrand to dotdigital Engagement Cloud.
This PR changes all the visible instances of the old "dotmailer Email Marketin" to "dotdigital Engagement Cloud" and the logo in the menu item.
I've tried to keep the changes as unaggressive as possible, so there are things that could also change (like the css class names) but I thought this could be safer.